requirements.txt

# Core dependencies 
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Deep Learning utilities
einops>=0.6.1
timm>=0.9.0

# Data processing
pandas>=2.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Pose estimation 
mediapipe>=0.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Evaluation metrics
lpips>=0.1.4
pytorch-fid>=0.3.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Demo
gradio>=3.50.0

# Progress and logging
tqdm>=4.65.0
wandb>=0.15.0

# Utilities
gdown>=4.7.0
requests>=2.31.0

# accelerate training
accelerate>=0.20.0

# OPTIONAL: DensePose / Detectron2

# human-body-prior
# detectron2 (install separately)
